server/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
httpx==0.28.1
openai==1.108.0
google-genai==1.31.0
python-dotenv==1.1.1
